refactor: move UserRead/UserCreate/UserUpdate to domain/auth/users
The schemata files should only contain ORM classes and create_db_and_tables.
Move the three pydantic request/response schemas out of schemata/user.py into
domain/auth/users.py, which is the natural home as it already owns all the
fastapi-users integration logic.

Update all import sites (routes/admin, routes/auth, routes/artifacts,
routes/plugins) to import from domain.auth.users.
