+++ backend/src/forecastbox/routes/plugins.py
@@ -21,10 +21,10 @@ from fastapi.responses import RedirectResponse, Response
 from fiab_core.fable import PluginCompositeId
 from pydantic import BaseModel
 
+from forecastbox.domain.auth.users import UserRead
 from forecastbox.domain.plugin.manager import PluginsStatus, modify_enabled, status_full, submit_update_single, uninstall_plugin
 from forecastbox.domain.plugin.store import PluginRemoteInfo, PluginStoreEntry, get_plugins_detail, submit_install_plugin
 from forecastbox.routes.admin import get_admin_user
-from forecastbox.schemata.user import UserRead
 from forecastbox.utility.config import config
 
 PREFIX = "/api/v1/plugin"
